sql: use `sql_parser_destroy()` to destroy parser
In this patch, all code related to parser destruction has been moved
to the `sql_parser_destroy()` function.

Part of #5485

NO_DOC=refactoring
NO_TEST=refactoring
NO_CHANGELOG=refactoring
